refactor utility classes and functions out of Builder view
